+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -1,3 +1,9 @@
+## v2.53.2
+
+* Deprecated `Promise#thenFinally()` - use `Promise#finally()`. The thenFinally
+   shim added to the promise module in v2.53.0 will be removed in v3.0
+   Sorry for the churn!
+
 ## v2.53.1
 
 * FIXED: for consistency with the other language bindings, `remote.FileDetector`